Store terminal failures as TerminalParseFailures
When attempting to print out a failure message using `failure_reason`
for my project `propose`[1], I was encountering the following error:

    NoMethodError: undefined method `unexpected' for [0, "[a-z]",
                   false]:Array> with backtrace

It looks like the problem here is that terminal failures are stored in
`@terminal_failures` as tuples, and when you call `terminal_failures` it
does a check to see if the first item is a `TerminalParseFailures`, and
converts the list if it isn't. However, if you call `terminal_failures`
multiple times, it's possible other errors were added, and so you'll
have a mixture of `Array` tuples and `TerminalParseFailure`s.

The fix is to store them as `TerminalParseFailure`s when the failure is
first encountered in `terminal_parse_failure`. It's not overly expensive
and allows us to avoid this unnecessary complexity.

[1] https://github.com/sds/propose