Provide nested parser class for Zeitwerk compat
Zeitwerk follows the general convention that the file name implies the constant
the file will load. Since the compiled Treetop files define two constant one
will not end up matching the filename, making it incompatible with reloaders
such as Zeitwerk.

To maintain compatibility with existing uses of Treetop we want to maintain
the <Grammar>Parser constant. But to enable compatibility with the convention
that Zeitwerk follows we also want a <Grammar>::Parser constant.

Therefore, the compiled code now defines both. First, it defines the nested
class within the grammar module. Then it aliases that constant to the existing
constant.

This should provide the best of both worlds. Users already using the old
constant can continue to do so. But users using auto-loaders like Zeitwerk
can now assume the standard naming conventions.

---

In addition to changing the output, I also rebuilt the Treetop gramar so the
new constant can be used on the included grammar. I also added a test to verify
the new nested class can be used as a parser.

---

This closes #51.