Add support for "endless" rule
This is similar to Ruby's "endless" method. The goal is to allow simple rules
to be more compact (similar also to JavaScript's bracket-less arrow functions).

Being more compact (while still readable) allows more code to be seen at once
and reduces keyword noise (lots of `end`s) and excess white space. These two
things make it easier to scan and grok the grammar.

This closes #54. See that issue for examples.