Fix race condition in constant removal from commit 62ca923
After 62ca923 intermittent issues are showing up in test suites with
NoMethodErrors. It looks like maybe there is a race condition with
removing/replacing constants. This patch attempts to fix that by
adding a rescue for the case where a constant has been successfully
removed. This should replace the old behavior where the constant is
replaced.